Final touch

+++ app/Http/Controllers/PostController.php
@@ -3,10 +3,20 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Storage;
 use App\Post;
 
 class PostController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index','show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -43,13 +53,30 @@ class PostController extends Controller
     {
         $this->validate($request,[
             'title'=>'required',
-            'body'=> 'required'
+            'body'=> 'required',
+            'cover_image'=>'image|nullable|max:1999'
         ]);
+        //handle file upload
+        if($request->hasFile('cover_image')){
+            //get filename with extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //Get just filename
+            $filename = pathinfo($filenameWithExt,PATHINFO_FILENAME);
+            //Get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }else{
+            $fileNameToStore = 'noimage.jpg';
+        }
         //create post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success','Post Created');
@@ -76,6 +103,9 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id != $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -92,10 +122,26 @@ class PostController extends Controller
             'title'=>'required',
             'body'=> 'required'
         ]);
+        //handle file upload
+        if($request->hasFile('cover_image')){
+            //get filename with extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //Get just filename
+            $filename = pathinfo($filenameWithExt,PATHINFO_FILENAME);
+            //Get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }
         //create post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success','Post Updated');
@@ -110,6 +156,13 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id != $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
+        if($post->cover_image != 'noimage.jpg'){
+            //delete image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
         return redirect('/posts')->with('success','Post Removed');
     }
